requirements-scheduler.txt
apache-beam[gcp]==2.59.0
pytest==7.2.2
jinja2-cli==0.8.2
pyyaml==6.0.2
